Fix vectorcall function signature for PyPy3.8.
Fixes https://github.com/cython/cython/issues/6867

+++ Cython/Utility/ObjectHandling.c
@@ -1833,7 +1833,7 @@ static PyObject *__Pyx_SelflessCall(PyObject *method, PyObject *args, PyObject *
 // See https://github.com/cython/cython/issues/6867
 static PyObject *__Pyx_SelflessCall(PyObject *method, PyObject **args, Py_ssize_t nargs, PyObject *kwnames) {
         return _PyObject_Vectorcall
-            (method, args ? args+1 : NULL, nargs ? (size_t) nargs-1 : 0, kwnames);
+            (method, args ? args+1 : NULL, nargs ? nargs-1 : 0, kwnames);
 }
 #else
 static PyObject *__Pyx_SelflessCall(PyObject *method, PyObject *const *args, Py_ssize_t nargs, PyObject *kwnames) {
